requirements.txt
alpaca-py
websockets
python-dotenv
